test-requirements.txt
pyyaml>=6.0.1
openapi_schema_validator
jsonpatch
jsonschema
